+++ src/components/BookDetail/BookDetail.tsx
@@ -8,15 +8,11 @@ import React, { useRef } from "react";
 import { useGetBookDetail } from "../../connection/LibraryQueryHooks";
 import { Book } from "../../model/Book";
 
-import { Button, Checkbox, Divider, FormControlLabel } from "@material-ui/core";
-import { Alert } from "@material-ui/lab";
-
+import { Divider } from "@material-ui/core";
 import { observer } from "mobx-react-lite";
-import { BookExtraPanels } from "./BookExtraPanels";
 import { LeftMetadata, RightMetadata } from "./MetadataGroup";
 import { DownloadsGroup } from "./DownloadsGroup";
 import { BookDetailHeaderGroup } from "./BookDetailHeaderGroup";
-import { DeleteButton } from "./DeleteButton";
 import { ReportButton } from "./ReportButton";
 import { Breadcrumbs } from "../Breadcrumbs";
 import { useTrack } from "../../analytics/Analytics";
@@ -48,6 +44,7 @@ import {
 } from "../appHosted/AppHostedUtils";
 import { Helmet } from "react-helmet";
 import { DownloadToBloomDialogs } from "./DownloadToBloomButton";
+import { BookOwnerControlsBox } from "./BookOwnerControlsBox";
 
 const BookDetail: React.FunctionComponent<IBookDetailProps> = (props) => {
     const l10n = useIntl();
@@ -100,11 +97,6 @@ const BookDetail: React.FunctionComponent<IBookDetailProps> = (props) => {
     }
 };
 
-// This should become true or just be removed once 5.7 is shipping.
-// The controls it hides require 5.7, so we don't want ordinary users to see them until then.
-// We do want to be able to test this on our dev site, though.
-const bloom57IsShipping = window.location.hostname.startsWith("dev");
-
 const BookDetailInternal: React.FunctionComponent<{
     book: Book;
     contextLangTag?: string;
@@ -126,8 +118,6 @@ const BookDetailInternal: React.FunctionComponent<{
     const embeddedMode = useIsEmbedded();
     const appHostedMode = useIsAppHosted();
     const user = LoggedInUser.current;
-    const userIsUploader = user?.username === props.book.uploader?.username;
-    const userIsModerator = useGetUserIsModerator();
     const l10n = useIntl();
     const getResponsiveChoice = useResponsiveChoice();
     const showDownloadDialog = useRef<() => void | undefined>();
@@ -267,7 +257,6 @@ const BookDetailInternal: React.FunctionComponent<{
                                         book={props.book}
                                         contextLangTag={props.contextLangTag}
                                     />
-                                    <DeleteButton book={props.book} />
                                 </div>
 
                                 {/* Enhance, maybe, add this and wire to some message <HowToPrintButton />*/}
@@ -320,55 +309,6 @@ const BookDetailInternal: React.FunctionComponent<{
                         )} */}
                             </div>
                         </div>
-                        {(user?.moderator || userIsUploader) && (
-                            <FormControlLabel
-                                css={css`
-                                    margin-top: 15px;
-                                    // By default, the checkbox has some padding used for animations on hover etc.
-                                    // And then, apparently a corresponding negative margin on this control makes it look
-                                    // aligned left. I think the amount of both is 8px. So a margin of -5px actually INCREASES
-                                    // the indent, aligning it with the 3px that something indents the Report button icon.
-                                    margin-left: -5px;
-                                `}
-                                control={
-                                    <Checkbox
-                                        css={css`
-                                            padding-top: 0;
-                                            margin-right: -5px;
-                                            padding-right: 1px;
-                                        `}
-                                        checked={props.book.draft}
-                                        onChange={(e) => {
-                                            props.book.draft = e.target.checked;
-                                            props.book.saveAdminDataToParse();
-                                        }}
-                                    />
-                                }
-                                label={
-                                    <div
-                                        css={css`
-                                            display: flex;
-                                        `}
-                                    >
-                                        <DraftIcon
-                                            css={css`
-                                                width: 54px;
-                                            `}
-                                        />
-                                        <div>
-                                            {l10n.formatMessage({
-                                                id:
-                                                    "book.detail.draftDescription",
-                                                defaultMessage:
-                                                    "Do not show this book to the public yet. I will share its URL with reviewers for feedback.",
-                                                description:
-                                                    "Label for a check box which, if checked, marks the book as 'DRAFT' and prevents the book from showing in most views",
-                                            })}
-                                        </div>
-                                    </div>
-                                }
-                            />
-                        )}
                         {props.book.draft && (
                             <DraftIcon
                                 css={css`
@@ -380,92 +320,13 @@ const BookDetailInternal: React.FunctionComponent<{
                                 `}
                             />
                         )}
-                        {(userIsUploader || userIsModerator) && (
-                            <div
-                                css={css`
-                                    display: flex;
-                                    flex-direction: column;
-                                `}
-                            >
-                                <h2
-                                    css={css`
-                                        margin-bottom: 0;
-                                        color: ${commonUI.colors.bloomBlue};
-                                    `}
-                                    id="book.detail.editing"
-                                >
-                                    Editing
-                                </h2>
-                                <Alert severity="info">
-                                    <div>
-                                        <FormattedMessage
-                                            id={"book.detail.updateBookNotice"}
-                                            defaultMessage={
-                                                "If you want to update this book with any changes, just upload it again from Bloom, using the same account. Your new version will replace this one."
-                                            }
-                                        />
-                                    </div>
-                                    {bloom57IsShipping && (
-                                        <div
-                                            css={css`
-                                                margin-top: 10px;
-                                            `}
-                                        >
-                                            <FormattedMessage
-                                                id={
-                                                    "book.detail.getForEditBookNotice"
-                                                }
-                                                defaultMessage={
-                                                    "If necessary, we can give you the book to edit in Bloom. You must first have Bloom 5.7 or greater installed ({downloadLink})."
-                                                }
-                                                values={{
-                                                    downloadLink: (
-                                                        <BlorgLink
-                                                            href="page/create/downloads"
-                                                            css={css`
-                                                                color: ${commonUI
-                                                                    .colors
-                                                                    .bloomBlue};
-                                                            `}
-                                                        >
-                                                            <FormattedMessage
-                                                                id={
-                                                                    "book.detail.downloadBloom"
-                                                                }
-                                                                defaultMessage={
-                                                                    "Download Bloom"
-                                                                }
-                                                            />
-                                                        </BlorgLink>
-                                                    ),
-                                                }}
-                                            />
-                                        </div>
-                                    )}
-                                </Alert>
-                                {bloom57IsShipping && (
-                                    <Button
-                                        onClick={() =>
-                                            showDownloadDialog.current?.()
-                                        }
-                                        color="secondary"
-                                        variant="outlined"
-                                        css={css`
-                                            align-self: flex-end;
-                                            margin-top: 5px;
-                                        `}
-                                    >
-                                        <FormattedMessage
-                                            id={"book.detail.editDownload"}
-                                            defaultMessage={
-                                                "Download into Bloom for editing"
-                                            }
-                                        />
-                                    </Button>
-                                )}
-                            </div>
+                        {user && (
+                            <BookOwnerControlsBox
+                                book={props.book}
+                                user={user}
+                                showDownloadDialog={showDownloadDialog}
+                            />
                         )}
-                        <BookExtraPanels book={props.book} />
                         <DownloadToBloomDialogs
                             book={props.book}
                             getShowFunction={(download) =>
