+++ src/localization/Code Strings.json
@@ -185,6 +185,10 @@
         "description": "This string is the heading of a page designed for a phone screen. It should fit across the width of such a screen. It labels a box where the user can type a partial language name and a page of matching language cards.",
         "message": "Language?"
     },
+    "book.detail.draft": {
+        "description": "",
+        "message": "Draft"
+    },
     "book.detail.editing": {
         "description": "",
         "message": "Editing"
@@ -275,6 +279,10 @@
     },
     "book.detail.editDownload": {
         "message": "Download into Bloom for editing",
+        "description": ""
+    },
+    "book.detail.youHavePermission": {
+        "message": "You have permission to modify this book",
         "description": ""
     },
     "book.metadata.level": {
